refactor(domain): refactors domain entities to unify and simplify base interfaces
